Move i18n to separate file
