+++ frontend/src/services/errorMessages.js
@@ -1,12 +1,13 @@
-import { i18n } from '../main';
+import i18n from './lang';
 
-export const splitCamelCase = (label) => label.replace(/([A-Z])/g, ' $1');
+export const splitCamelCase = label =>
+  label.replace(/([a-z])([A-Z])/g, '$1 $2');
 
 /*
  * Compose error message for invalid field in user language.
  * TODO: Test
  */
-export const renderErrorMessage = (field) =>
+export const renderErrorMessage = field =>
   i18n.t('drive_form.validation_error', {
     field: splitCamelCase(field),
   });
